+++ beltradar/tasks.py
@@ -9,7 +9,7 @@ from allianceauth.services.tasks import QueueOnce
 
 # AA Belt Radar
 from beltradar import __title__, app_settings
-from beltradar.models import BeltSurveySession
+from beltradar.models import BeltSurveySession, EveMarketPrice
 from beltradar.providers import AppLogger
 
 logger = AppLogger(get_extension_logger(__name__), __title__)
@@ -31,9 +31,9 @@ TASK_DEFAULTS_BIND_ONCE = {**TASK_DEFAULTS, **{"bind": True, "base": QueueOnce}}
 # Default params for tasks that need run once only.
 TASK_DEFAULTS_ONCE = {**TASK_DEFAULTS, **{"base": QueueOnce}}
 
-TASK_DEFAULTS_BIND_ONCE_OWNER = {
+TASK_DEFAULTS_BIND_ONCE_SESSION = {
     **TASK_DEFAULTS_BIND_ONCE,
-    **{"once": {"keys": ["eve_id"], "graceful": True}},
+    **{"once": {"keys": ["public_id"], "graceful": True}},
 }
 
 
@@ -46,12 +46,18 @@ def update_all_belt_radar(runs: int = 0, force_refresh: bool = False):
             kwargs={"public_id": session.public_id, "force_refresh": force_refresh}
         )
         runs = runs + 1
-
     logger.info("Queued %s Session Tasks", runs)
 
 
 # pylint: disable=unused-argument
-@shared_task(**TASK_DEFAULTS_BIND_ONCE_OWNER)
+@shared_task(**TASK_DEFAULTS_BIND_ONCE_SESSION)
 def update_belt_radar_session(self, public_id: str, force_refresh: bool = False):
     """Update a single belt radar session."""
     # Placeholder for future implementation, currently no background processing needed for session updates.
+
+
+@shared_task(**TASK_DEFAULTS_ONCE)
+def update_market_prices():
+    """Update market prices for all market items."""
+    count = EveMarketPrice.objects.update_from_esi()
+    logger.info("Updated market prices for %s items.", count)